+++ Classes/BannerSampleScene.h
@@ -4,7 +4,11 @@
 #include "cocos2d.h"
 #include "bidmad/BannerInterface.h"
 
+#if (CC_TARGET_PLATFORM == CC_PLATFORM_IOS)
+class BannerSampleScene : public cocos2d::Scene
+#elif (CC_TARGET_PLATFORM == CC_PLATFORM_ANDROID)
 class BannerSampleScene : public cocos2d::Scene, Application
+#endif
 {
 public:
     static cocos2d::Scene* createScene();
@@ -23,9 +27,12 @@ public:
     void backScene();
 
     // implement the "static create()" method manually
+#if (CC_TARGET_PLATFORM == CC_PLATFORM_IOS)
+#elif (CC_TARGET_PLATFORM == CC_PLATFORM_ANDROID)
     virtual bool applicationDidFinishLaunching();
     virtual void applicationDidEnterBackground();
     virtual void applicationWillEnterForeground();
+#endif
     CREATE_FUNC(BannerSampleScene);
 };