+++ Classes/bidmad/BannerInterface.h
@@ -1,5 +1,6 @@
 #include "cocos2d.h"
 #if (CC_TARGET_PLATFORM == CC_PLATFORM_IOS)
+#include "ios/BannerBridgeCpp.h"
 #elif (CC_TARGET_PLATFORM == CC_PLATFORM_ANDROID)
 #include "android/BannerController.h"
 #endif
@@ -9,6 +10,8 @@ USING_NS_CC;
 class BannerInterface{
 private:
 #if (CC_TARGET_PLATFORM == CC_PLATFORM_IOS)
+    BannerBridgeCpp* banner;
+    char* localZoneId;
 #elif (CC_TARGET_PLATFORM == CC_PLATFORM_ANDROID)
     BannerController* banner;
 #endif
@@ -21,8 +24,12 @@ public:
     void removeBanner();
     void hideBannerView();
     void showBannerView();
+    
+#if (CC_TARGET_PLATFORM == CC_PLATFORM_IOS)
+#elif (CC_TARGET_PLATFORM == CC_PLATFORM_ANDROID)
     void onPause();
     void onResume();
+#endif
 
     void setOnLoadCallback(void (*_onLoadCallback) (char *));
     void setOnFailCallback(void (*_onFailCallback) (char *));